refactor: reuse utils
